backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic-settings
motor
python-whois
dnspython
